Refactor refresh_pairs to exchange and fix tests
