+++ freqtrade/tests/test_freqtradebot.py
@@ -44,7 +44,7 @@ def patch_get_signal(freqtrade: FreqtradeBot, value=(True, False)) -> None:
     """
     freqtrade.strategy.get_signal = lambda e, s, t: value
     freqtrade.exchange.get_candle_history = lambda p, i: None
-    freqtrade.refresh_tickers = lambda i: True
+    freqtrade.exchange.refresh_tickers = lambda p, i: True
 
 
 def patch_RPCManager(mocker) -> MagicMock:
@@ -135,36 +135,6 @@ def test_throttle_with_assets(mocker, default_conf) -> None:
 
     result = freqtrade._throttle(throttled_func, min_secs=0.1)
     assert result == -1
-
-
-def test_refresh_tickers(mocker, default_conf, caplog) -> None:
-    tick = [
-         [
-             1511686200000,  # unix timestamp ms
-             1,  # open
-             2,  # high
-             3,  # low
-             4,  # close
-             5,  # volume (in quote currency)
-         ]
-     ]
-
-    async def async_get_candles_history(pairlist, timeframe):
-        return [(pair, tick) for pair in pairlist]
-
-    caplog.set_level(logging.DEBUG)
-    freqtrade = get_patched_freqtradebot(mocker, default_conf)
-    freqtrade.exchange.async_get_candles_history = async_get_candles_history
-
-    pairs = ['IOTA/ETH', 'XRP/ETH']
-    # empty dicts
-    assert not freqtrade.exchange.klines
-    freqtrade.refresh_tickers(['IOTA/ETH', 'XRP/ETH'])
-
-    assert log_has(f'Refreshing klines for {len(pairs)} pairs', caplog.record_tuples)
-    assert freqtrade.exchange.klines
-    for pair in pairs:
-        assert freqtrade.exchange.klines[pair]
 
 
 def test_gen_pair_whitelist(mocker, default_conf, tickers) -> None:
